Refactor; changed Decimal to float

+++ apps/backend/src/modules/currency/adapters/api/dto/incoming/calculate_currency_dto.py
@@ -1,4 +1,3 @@
-from decimal import Decimal
 
 from pydantic import Field
 
@@ -9,4 +8,4 @@ from shared.models.camel_model import CamelModel
 class IncomingCalculateCurrencyDTO(CamelModel):
     currency_from: CurrencyEnum = Field(..., description="Base currency you want to exchange." )
     currency_to: CurrencyEnum = Field(..., description="Target currency." )
-    value: Decimal = Field(..., description="Value of base currency." )
+    value: float = Field(..., description="Value of base currency." )
